ZameenScrapingAPI/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Flask_Cors==4.0.1
python_dateutil==2.9.0
Requests==2.32.3
